+++ lib/Indexer.php
@@ -2,10 +2,12 @@
 
 namespace SearchEngine;
 
+use ProcessWire\HookEvent;
+
 /**
  * SearchEngine Indexer
  *
- * @version 0.11.0
+ * @version 0.12.0
  * @author Teppo Koivula <teppo.koivula@gmail.com>
  * @license Mozilla Public License v2.0 http://mozilla.org/MPL/2.0/
  */
@@ -24,6 +26,20 @@ class Indexer extends Base {
      * @var Processor
      */
     protected $processor;
+
+    /**
+     * Module specific extra actions
+     *
+     * @var array
+     */
+    protected $moduleExtras = [];
+
+    /**
+     * Local instance of FormBuilder module
+     *
+     * @var \ProcessWire\FormBuilder|null
+     */
+    protected $formBuilder = null;
 
     /**
      * Constructor method
@@ -89,6 +105,7 @@ class Indexer extends Base {
             $save = false;
         }
         if ($page->id && ($return == 'index' || $index_field_exists)) {
+            $this->initModuleExtras();
             if ($this->wire('modules')->isInstalled('LanguageSupport') && $this->wire('fields')->get($index_field)->type == 'FieldtypeTextareaLanguage') {
                 foreach ($this->wire('languages') as $language) {
                     $index[$language->id] = $this->getPageIndex($page, $options['indexed_fields'], '', [
@@ -299,6 +316,49 @@ class Indexer extends Base {
             }
         }
         return $index;
+    }
+
+    /**
+     * Init module extras
+     */
+    protected function initModuleExtras() {
+        if (!isset($this->moduleExtras['FormBuilder']) && $this->wire('modules')->isInstalled('FormBuilder')) {
+            $this->moduleExtras['FormBuilder'] = in_array('FormBuilder', $this->getOptions()['module_extras'] ?? []);
+            if ($this->moduleExtras['FormBuilder']) {
+                $this->addHookAfter('FieldtypeTextarea::formatValue', $this, 'embedFormBuilder');
+            }
+        }
+    }
+
+    /**
+     * Embed FormBuilder forms
+     *
+     * This method looks for FormBuilder embed tag(s) within text content and attempts to replace them with rendered
+     * markup of said form(s) in order to make actual form content searchable.
+     *
+     * @param HookEvent $event
+     */
+    protected function embedFormBuilder(HookEvent $event) {
+        $field = $event->arguments[1];
+        if ($this->formBuilder === null) {
+            $this->formBuilder = $this->wire('modules')->get('FormBuilder');
+        }
+        if (!in_array($field->id, $this->formBuilder->embedFields) || strpos($event->return, ">" . $this->formBuilder->embedTag . "/") === false) {
+            // field cannot be used for embedding forms or no embed codes detected, bail out early
+            return;
+        }
+        if (preg_match_all('!<([^>]+)>' . $this->formBuilder->embedTag . '/([-_a-zA-z0-9]+)\s*</\\1>!', $event->return, $matches)) {
+            foreach ($matches[0] as $key => $tag) {
+                try {
+                    $this->formBuilderRender = $this->formBuilder->render($matches[2][$key]);
+                    if ($this->formBuilderRender instanceof \ProcessWire\FormBuilderRender) {
+                        $event->return = str_replace($tag, $this->formBuilderRender->render(), $event->return);
+                    }
+                } catch (\Exception $e) {
+                    // rendering the form failed, not much we can do about that now
+                }
+            }
+        }
     }
 
 }